Changed PlayerPhysicsComponent to update based on time instead of frame rate, improved performance a little bit. Also found out that fullscreen mode is MUCH faster than windowed mode, but the display is very weird, will look more into this later

+++ include/GameEntity.h
@@ -27,7 +27,7 @@ class GameEntity
 
     public:
         //Fields
-        int xPos, yPos;
+        float xPos, yPos;
 
         //Constructor
         GameEntity();
